requirement.txt
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.19.5
Pillow>=8.0.0
matplotlib>=3.3.4
opencv-python>=4.5.3
tensorboard>=2.4.1
rawpy>=0.16.0
imageio>=2.9.0
tqdm>=4.62.0
dlib>=19.22.0